requirements.txt
pyspark==3.5.2
kafka-python==2.0.2
